+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This file lists the changes that have occurred since January 2024 in the project
 
 ## Unreleased
 
+## 0.1 - 2024-08-21
+
 * Added `crusader remote` command to start a web server listening on port 35482.
    It allows starting tests on a separate machine and displays the resulting charts in the web page.
 * Use system fonts in GUI